requirements.txt
pyrsistent
matplotlib
python-decouple
dj-database-url
dj-static
gunicorn == 20.1.0
suds==1.0.0
requests==2.2.0

drf-yasg==1.20.0

#Django
Django==4.0.2
django-cleanup==7.0.0
django-crum==0.7.9
django-environ==0.10.0
django-user-agents==0.4.0
django-simple-history==3.0.0
django-cors-headers==4.3.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0

#MySQL
mysqlclient
PyMySQL==1.0.2
django-mysql==2.2.0
asgiref==3.6.0
Brotli==1.0.9
cffi==1.15.1
cssselect2==0.7.0
fonttools==4.38.0
html5lib==1.1
Pillow==9.4.0
psycopg2-binary==2.9.6
pycparser==2.21
pydyf==0.5.0
pyphen==0.13.2
six==1.16.0
sqlparse==0.4.3
tinycss2==1.2.1
tzdata==2022.7
ua-parser==0.16.1
user-agents==2.2.0
weasyprint==58.0
webencodings==0.5.1
zopfli==0.2.2

#LIBRERIAS GENERAR GRAFICO DE MODELOS DJANGO
django-extensions==3.2.3
pyparsing==2.4.7
pydot==1.4.2